move all libs into main folider
