+++ cnc_ctrl_v1/FormatDouble.cpp
