+++ cnc_ctrl_v1/bcconfig.h
