+++ cnc_ctrl_v1/gpl.txt
